# Conflicts:
#	ms-api/src/main/java/com/hq/ecmp/ms/api/ScheduledJobs/ScheduledTask.java
#	ms-api/src/main/java/com/hq/ecmp/ms/api/controller/account/InvoiceInfoController.java
#	ms-api/src/main/java/com/hq/ecmp/ms/api/controller/city/CityController.java
#	ms-api/src/main/java/com/hq/ecmp/ms/api/controller/order/OrderController.java
#	ms-api/src/main/resources/config/mybatis/InvoiceInfoMapper.xml
#	ms-api/src/main/resources/config/mybatis/OrderInfoMapper.xml
#	ms-api/src/test/java/com/hq/ecmp/ms/api/controller/base/OrgControllerTest.java
#	ms-core/src/main/java/com/hq/ecmp/mscore/domain/CarInfo.java
#	ms-core/src/main/java/com/hq/ecmp/mscore/domain/InvoiceInfo.java
#	ms-core/src/main/java/com/hq/ecmp/mscore/dto/InvoiceByTimeStateDTO.java
#	ms-core/src/main/java/com/hq/ecmp/mscore/dto/InvoiceInsertDTO.java
#	ms-core/src/main/java/com/hq/ecmp/mscore/mapper/OrderInfoMapper.java
#	ms-core/src/main/java/com/hq/ecmp/mscore/service/IDriverInfoService.java
#	ms-core/src/main/java/com/hq/ecmp/mscore/service/impl/DriverInfoServiceImpl.java
#	ms-core/src/main/java/com/hq/ecmp/mscore/service/impl/EcmpUserServiceImpl.java
#	ms-core/src/main/java/com/hq/ecmp/mscore/service/impl/OrderInfoServiceImpl.java

+++ ms-core/src/main/java/com/hq/ecmp/mscore/dto/cost/CostConfigListResult.java
@@ -6,6 +6,7 @@ import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -40,19 +41,19 @@ public class CostConfigListResult {
     @ApiModelProperty("车型集合")
     private List<CostConfigCarTypeInfo> carTypes;
     @ApiModelProperty(value = "起步价，非包车类型使用")
-    private Double startPrice;
+    private BigDecimal startPrice;
     @ApiModelProperty(value = "套餐价，包车类型使用")
-    private Double combosPrice;
+    private BigDecimal combosPrice;
     @ApiModelProperty(value = "套餐里程（：公里）")
-    private Double combosMileage;
+    private BigDecimal combosMileage;
     @ApiModelProperty(value = "套餐时长（：分钟）")
     private Long combosTimes;
     @ApiModelProperty(value = "超里程单价（元/公里）")
-    private Double beyondPriceEveryKm;
+    private BigDecimal beyondPriceEveryKm;
     @ApiModelProperty(value = "超里程时长（元/分钟）")
-    private Double beyondPriceEveryMinute;
+    private BigDecimal beyondPriceEveryMinute;
     @ApiModelProperty(value = "等待费（元/分钟）")
-    private Double waitPriceEreryMinute;
+    private BigDecimal waitPriceEreryMinute;
     @ApiModelProperty(value = "创建时间")
     @JsonFormat(pattern = "yyyy-MM-dd",timezone ="GMT+8" )
     private Date createTime;
